Add best subsonic client post!
I guess this is the first time I used strikethrough because I just
learned that Markdown doesn't support it! We install an extra dependency
`pymdown-extensions` to give us support for that. While I was at it I
just updated the rest of the dependencies.

In the future I'd like to figure out a way to scope CSS and put output
files into folders, but I'm not worried about it now.