Create rss feed
This was surprisingly easy! It's mostly based on [the spec], which was
pretty readable. The hardest part was getting the RFC 822 time correct
with the local time zone. I had to add timezones to my posts, which
wasn't too hard because there weren't that many, but accounting for
Daylight Savings Time was kinda lame. I think I did them right, but it's
not that important if I'm an hour off anyways.

[the spec]: https://www.rssboard.org/rss-specification